contrib: added previous implementation of some helper scripts
Those are the remainder of collection of python scripts, initially
implemented in context of my NixOS configuration [1]/[2], and the
most important part of that collection is already implemented here.

This remainder is yet to be reimplemented, the way and language I
see fit. The usage context of these scripts, that may be useful
for reimplementation, could be seen at [3]/[4].

[1] https://github.com/wiedzmin/nixos-config
[2] [[orgit:~/workspace/repos/github.com/wiedzmin/nixos-config/][~/workspace/repos/github.com/wiedzmin/nixos-config/ (magit-status)]]
[3] https://github.com/wiedzmin/nixos-config/commit/4e0b0f148efd5a61215c6ac648e0a8061f5f5e02
[4] [[orgit-rev:~/workspace/repos/github.com/wiedzmin/nixos-config/::4e0b0f148efd5a61215c6ac648e0a8061f5f5e02][~/workspace/repos/github.com/wiedzmin/nixos-config/ (magit-rev 4e0b0f14)]]